BackEnd/requirements.txt
Flask>=3.0.0
Flask_Cors>=4.0.0
Flask_Login>=0.6.3
flask_marshmallow>=0.15.0
flask_sqlalchemy>=3.1.1
forex_python>=1.8
python-dotenv>=1.0.0
Werkzeug>=3.0.1
psycopg2-binary